add BusinessesListFragment.kt to desplay businesses in recyclerView and add search bar and update design

+++ app/src/main/java/com/bignerdranch/android/yelpsbusinessesweather/model/Weather.kt
@@ -26,7 +26,7 @@ data class Forecast(
 )
 data class Forecastday(
     val date: String,
-    val date_epoch: Int,
+    val date_epoch: Long,
     val day: Day,
     val hour: List<Hour>
 )
